+++ src/models/incidentals/ongoing_incidentals.ts
@@ -14,7 +14,22 @@ export interface OngoingIncidentalsCost {
   /**
    * Month in which the cost first applies
    */
-  date: MonthYear;
+  dueDate: MonthYear;
+
+  /**
+   * Date the bill was paid
+   */
+  paymentDate: Date;
+
+  /**
+   * Document id of the bill
+   */
+  billDocumentId?: string;
+
+  /**
+   * Document id of the bank transfer
+   */
+  bankTransferDocumentId: string;
 }
 
 /**