+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - One time incidentals
     - Bill document
     - Bank transfer document
+- Ongoing incidentals
+    - Bill document
+    - Bank transfer document
+    - Payment date
 
 ### Changed